+++ cpp/onedal/decomposition/pca/common.cpp
@@ -29,7 +29,7 @@ using detail::model_impl;
 params_base::params_base()
   : impl_(new params_impl()) { }
 
-std::int32_t params_base::get_components_count() const {
+std::int64_t params_base::get_components_count() const {
   return impl_->components_count;
 }
 
@@ -37,7 +37,7 @@ bool params_base::get_is_deterministic() const {
   return impl_->is_deterministic;
 }
 
-void params_base::set_components_count(std::int32_t value) {
+void params_base::set_components_count(std::int64_t value) {
   impl_->components_count = value;
 }
