+++ cpp/onedal/decomposition/pca/common.hpp
@@ -44,10 +44,10 @@ class params_base : public base {
 
   params_base();
 
-  std::int32_t get_components_count() const;
+  std::int64_t get_components_count() const;
   bool get_is_deterministic() const;
 
-  void set_components_count(std::int32_t value);
+  void set_components_count(std::int64_t value);
   void set_is_deterministic(bool value);
 
  private:
@@ -61,7 +61,7 @@ class params : public params_base {
   using float_t  = Float;
   using method_t = Method;
 
-  auto& set_components_count(int32_t value) {
+  auto& set_components_count(int64_t value) {
     params_base::set_components_count(value);
     return *this;
   }
